Add inital offline install capabilities (#1121)
Supersedes #985
Fixes #984

+++ functions/openhabian.bash
@@ -25,6 +25,7 @@ get_git_revision() {
 install_cleanup() {
   echo -n "$(timestamp) [openHABian] Cleaning up... "
   if ! cond_redirect systemctl -q daemon-reload &> /dev/null; then echo "FAILED (daemon-reload)"; return 1; fi
+  if ! cond_redirect apt-get clean; then echo "FAILED (apt-get clean)"; return 1; fi
   if cond_redirect apt-get autoremove --yes; then echo "OK"; else echo "FAILED"; return 1; fi
 }
